requirements.txt
flask
sqlalchemy
python-docx
apscheduler
psycopg2-binary
